Move util to pkg/duck
